Move AvatarRow to designsystem module, because it will be used by spaces.
